% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_column.R
\name{split_column}
\alias{split_column}
\title{Split a Column into Multiple Columns}
\usage{
split_column(data, column, into, sep = " ", remove_original = TRUE)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{column}{Column name to split (character).}

\item{into}{Character vector of new column names.}

\item{sep}{Separator character or regular expression.}

\item{remove_original}{Logical. Remove the original column? Default: TRUE.}
}
\value{
Tibble with new columns.
}
\description{
Splits a string column into several columns using a delimiter.
}
\examples{
df <- tibble::tibble(name = c("John Smith", "Jane Doe"))
split_column(df, column = "name", into = c("first", "last"), sep = " ")

}
